+++ app/javascript/components/Snackbar/index.js
