+++ app/javascript/components/Task/Task.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Card, CardHeader, CardContent, Typography, IconButton } from '@material-ui/core';
 import { Edit } from '@material-ui/icons';
 import useStyles from './useStyles';
-import { dateHelper } from '../../../helpers';
+import { dateHelper } from '../../helpers';
 
 function Task({ task, onClick }) {
   const styles = useStyles();
